+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsejsonschema.py
@@ -40,9 +40,9 @@ class EvaluatorResponseJSONSchema(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-29T07:30:32.803Z"
+    created: Optional[str] = "2026-07-29T10:16:55.110Z"
 
-    updated: Optional[str] = "2026-07-29T07:30:32.803Z"
+    updated: Optional[str] = "2026-07-29T10:16:55.110Z"
 
     updated_by_id: OptionalNullable[str] = UNSET