+++ packages/orq-rc/RELEASES.md
@@ -9809,3 +9809,13 @@ Based on:
 - [python v4.13.0-rc.13] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.13] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.13 - packages/orq-rc
+
+## 2026-07-29 10:20:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.14] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.14 - packages/orq-rc